+++ Tests/Realm.Tests/Database/RealmObjectTests.cs
@@ -28,8 +28,6 @@ using NUnit.Framework;
 using Realms.Exceptions;
 using Realms.Schema;
 #if TEST_WEAVER
-using TestAsymmetricObject = Realms.AsymmetricObject;
-using TestEmbeddedObject = Realms.EmbeddedObject;
 using TestRealmObject = Realms.RealmObject;
 #else
 using TestRealmObject = Realms.IRealmObject;